+++ src/ppux.h
@@ -1,3 +1,3 @@
 #include "fundude.h"
 
-void ppu_step(fundude* fd, int cycles);
+void ppu_step(fundude* fd, uint8_t cycles);
